refactor(auth): move createSession to shared util; rename user lookup
- Move createSession from the sign-in route folder to app/utils/session.server.ts
  so shared utilities no longer import from a route module (fixes inverted
  layering; unblocks reuse by other sign-in methods).
- Rename findActiveUserByEmail to findExistingUserByEmail: the User model has no
  active/enabled state, so it is purely an existence check. (per PR review)
